Only fetch data from server once a week
Use locally cached data instead of going to the server most of the time.   If the cache is more than 7 days old fetch data from the server.  Because a list of years that we have scheduled data is now available remove the checkRides function.

Year selection code commented out for now.