Link to maps app
Turn the route item on the ride detail page into a link that will bring you to maps with a route from your current location to the restaurant.  You can follow the route or go back to the app.  Needs testing to make sure the maps URL create is correct for all of our restaurants.  This will be part of version 1.1.  Thanks to Carl T for the idea.