Use a button instead of an onTapGesture
Create a circular button style for the SMBC riders image.  Use the image in a button of that style to go directly to the next ride.

I suspect there is a better way than creating two navigation links with an EmptyView() that are triggered by a selection, but I haven't found it yet.