Move MapInfoView to a separate file
Also try to get map display to change when a new map type is selected.  It kind of works but I'm running into either a bug in my code or a bug in swiftui where when the view is invoked from a list updating the map returns to the list.