Add help button
Update UI test to check help button, too.

The text of the help button give simple instructions that pertain mostly to the ability to tap on the center image to go to the next ride and/or log press on the image to force schedule data reload from the SMBC server.