Map Marker Selection
Prep for showing a lookaround view when a restaurant marker is selected.  However, although the marker is highlighted as if it was selected the selectedItem binding is not changed and onChange(of:) is not called.  I don't know if I am doing something wrong or if this is a SwiftUI Map beta issue.