Swift 6 at project level plus explicit modules
Branch also renamed.  This is for Xcode16, but still iOS 17.  There is a
bug with the markdown package which has been tracked down. The bug seems
to be in the compiler but the workaround in the markdown code (make one
of the enums explicitly conform to sendable) is probably correct and I
hope the fix will be pulled, soon.