Bug fix: race causing out of sync ride ids
The ride id was a UUID created whenever a ScheduledRide was init.  If a
ride refresh occured after showing the current ride a new table of rides
was created with different UUIDs than the ride showing on the screen.
Back forward swipes wouldn't work because the ride couldn't be found.

Change the ID to be calculated based upon the start data and either the
restaurant code or the trip code (first word of description).