Schedule Reducer Testing and fallout
Testing is good.  I realized that the loadInProgress variable needs to
be more than a boolean to distinguish between and initial load and a
duplicate request. ScheduleReducer changed.

The schedule tests have to run sequentially or one test might removes a
cache file another test depends upon. An xctestplan added to tell xcode
to run all schedule tests in sequence.